refactor(test): 인수 테스트 리팩터링
